Properly queue runners so worker does not get killed on first runner completing
